+++ napp/src/ViewContainer.js
@@ -1,123 +1,41 @@
 define([
     "dojo/_base/declare",
-    "dojo/_base/lang",
-    "dojo/promise/all",
-    "dojo/Deferred",
-    "dojo/hash",
-    "dojo/router",
-    "dojo/when",
     
-    "dijit/_WidgetBase",
+    "dijit/layout/_LayoutWidget",
     "dijit/_TemplatedMixin",
     "dijit/_WidgetsInTemplateMixin",
     
     "dijit/layout/BorderContainer",
     "dijit/layout/ContentPane",
 
-    "napp/utils/lang",
-
     "dojo/text!./templates/ViewContainer.html"
 ], function (
     declare,
-    lang,
-    all,
-    Deferred,
-    hash,
-    router,
-    when,
     
-    _WidgetBase, 
+    _LayoutWidget, 
     _TemplatedMixin, 
     _WidgetsInTemplateMixin,
     
     BorderContainer,
     ContentPane,
 
-    langUtil,
 
     template) {
     
-    return declare([_WidgetBase, _TemplatedMixin, _WidgetsInTemplateMixin], {
+    return declare([_LayoutWidget, _TemplatedMixin, _WidgetsInTemplateMixin], {
         templateString: template,
-        
-        createViewContext: null, // Injected
-        viewsSpec: null, // Injected
-        defaultView: null, // Injected
 
-        _viewSpecsHash: null,        
-        _currentViewWidget: null,
-
-        postCreate: function () {
-            this.inherited(arguments);
-
-            router.register(/^(\w*)(\?*)(.*)$/, (this, function (evt) {
-                var viewName = evt.params[0],
-                    paramString = evt.params[1],
-                    paramHash = {};
-                
-                var regex = /([^=]+)=([^=]+)(?:&|$)/g;
-                var match;
-                
-                while ((match = regex.exec(paramString)) !== null) {
-                    paramHash[match[1]] = match[2];
-                }
-                                
-                this._loadView(viewName, paramHash);
-            }).bind(this));
-
-            // TODO: Extract to view loader component
-            var viewSpecLoaders = this.viewsSpec.map(function (spec) {
-                if (spec.constructor == String) {
-                    var d = new Deferred();
-                    require ([spec], function (loadedSpec) {
-                        d.resolve(loadedSpec);
-                    });
-                    return d;
-                } else if (spec.constructor == Object) {
-                    return spec;
-                }
-
-                return null;
-            });
-
-            when (all(viewSpecLoaders), function (viewsSpecs) {
-                this._viewSpecsHash = viewsSpecs.reduce(function (accumulated, item) {
-                    return langUtil.deepMixin(accumulated, item);
-                }, {});                
-
-                router.startup();
-                if (hash() === "") {
-                    router.go(this.defaultView);
-                }
-            }.bind(this));
+        addChild: function (widget) {
+            widget.region = "center";
+            this._container.addChild(widget);
         },
-        
-        destroy: function () {
-            this._unloadCurrentView();
-            this.inherited(arguments);
+
+        removeChild: function (widget) {
+            this._container.removeChild(widget);
         },
-        
-        _loadView: function (viewName, params) {
-            var specToLoad = {};
-            specToLoad[viewName] = this._viewSpecsHash[viewName];
 
-            when(this.createViewContext(specToLoad), function (loadedViewSpec) {
-                this._unloadCurrentView();
-
-                widget = loadedViewSpec[viewName].view;
-                widget.region = "center";
-                this._container.addChild(widget);
-
-                this._currentViewWidget = widget;                
-            }.bind(this));
-        },
-        
-        _unloadCurrentView: function () {
-            if (this._currentViewWidget) {
-                this._container.removeChild(this._currentViewWidget);
-                this._currentViewWidget.destroy();
-                this._currentViewWidget = null;
-            }      
+        layout: function () {
+            this._container.resize();
         }
     });
 });
